requirements.txt
pyqt5
pytest
pyyaml